refactor: implement user type
