+++ lib/prisma/utils.ts
@@ -1,8 +1,8 @@
-import { UserId } from '@/types/next-auth';
+import { User, UserId } from '@/types/user';
 import { prismaClient } from '@/lib/prisma';
 
-export async function getUser(userId: UserId) {
-  return prismaClient.user.findUnique({
+export async function getUser(userId: UserId): Promise<User | null> {
+  const user = await prismaClient.user.findUnique({
     select: {
       id: true,
       name: true,
@@ -12,4 +12,6 @@ export async function getUser(userId: UserId) {
     },
     where: { id: userId },
   });
+
+  return (user as User) ?? null;
 }
