+++ types/next-auth.d.ts
@@ -1,9 +1,6 @@
 import { DefaultSession, Session, User } from 'next-auth';
 import { JWT } from 'next-auth/jwt';
-import { prismaClient } from '@/lib/prisma';
-
-type UserId = PrismaClient.user['id'];
-type Role = 'user' | 'admin';
+import { Role, UserId } from '@/types/user';
 
 declare module 'next-auth' {
   interface Session {
